Fix bugs by loading vim-sensible first
In particular I was having an issue where vim-fugitive would load before
vim-sensible. This meant that `filetype on` was running late in the
process and it's autocmd was fairly late in the stack. This meant that
when the ruby ftplugin was re-initialized it would reset variables (the
tags option in particular thanks to the `b:undo_ftplugin` setting) but
the earlier plugins would not get a chance to reset them.